+++ app/routes/components/Button/index.tsx
@@ -3,8 +3,8 @@ import { Section } from "src/components/Section";
 import { Typography } from "@/components/Typography";
 import type { Route } from "./+types/index";
 import { changelog } from "./changelog.codegen";
-import { Demo } from "./Demo";
-import { demoSourceCode } from "./Demo_source_code.codegen";
+import Demo01 from "./demos/Demo01";
+import demo01sourceCode from "./demos/Demo01.source_code.codegen";
 import { sourceCode } from "./source_code.codegen";
 
 export function meta(_: Route.MetaArgs) {
@@ -18,7 +18,7 @@ export default function ButtonPage() {
   return (
     <ComponentTemplate component="Button">
       <Section.Root title="Button">
-        <Section.Demo node={<Demo />} code={demoSourceCode} />
+        <Section.Demo node={<Demo01 />} code={demo01sourceCode} />
         <Typography.H2 mt="1.5rem" mb="1rem">
           Usage
         </Typography.H2>