+++ src/components/Button/CHANGELOG.md
@@ -10,3 +10,4 @@
   from ThemeProvider.
 - 2025-10-04: Add `ghost` variant. Code formatting improvement.
 - 2025-10-08: Add `light` variant, update `ghost` variant styles.
+- 2025-10-15: Support change styles when `disabled`.